src/tools/evm_tools.py
from typing import Any, Dict, List, Optional
from web3 import Web3
from langchain_core.tools import tool

class EVMTools:
    def __init__(self, RPC_URL: str):
        self.web3 = Web3(Web3.HTTPProvider(RPC_URL))
        
    @tool
    def get_contract_abi(self, contract_address: str) -> Dict[str, Any]:
        """Get the ABI of a contract from Etherscan."""
        # TODO: Implement Etherscan API integration
        raise NotImplementedError("Etherscan API integration not implemented yet")
    
    @tool
    def get_contract_functions(self, contract_address: str) -> List[str]:
        """Get all available functions of a contract."""
        # TODO: Implement contract function listing
        raise NotImplementedError("Contract function listing not implemented yet")
    
    @tool
    def call_contract_function(self, contract_address: str, function_name: str, *args) -> Any:
        """Call a contract function with given arguments."""
        # TODO: Implement contract function calling
        raise NotImplementedError("Contract function calling not implemented yet")
    
    @tool
    def get_contract_balance(self, contract_address: str) -> int:
        """Get the ETH balance of a contract."""
        return self.web3.eth.get_balance(contract_address) 