requirements.txt
langgraph-codeact>=0.1.0
langchain>=0.1.0
langchain-anthropic>=0.0.1
web3>=6.11.1
python-dotenv>=1.0.0 